Создана папка с башнями. Вынесен в отдельный файл класс BaseTower
